changed decimal to float

+++ backend/db/models/spot.js
@@ -55,12 +55,12 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat:{
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       allowNull: false
 
     },
     lng:{
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       allowNull: false
     },
     name: {
@@ -68,7 +68,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     description: DataTypes.STRING,
-    price: DataTypes.DECIMAL
+    price: DataTypes.FLOAT
   }, {
     sequelize,
     modelName: 'Spot',
